pkg/jsonmessage: should return early on context cancellation
This patch makes the `jsonmessage.DisplayJSONMessagesToStream` and
`jsonmessage.DisplayJSONMessagesStream` context aware and allows
them to return early upon receiving a context.Cancellation.

The current behavior of `jsonmessage.DisplayJSONMessagesToStream` and
`jsonmessage.DisplayJSONMessagesStream` are to continue to read the
response object for as long as the reponse remains open.

Since these functions do not return an error upon closing the response
object it is impossible to know if the `jsonmessage.DisplayJSONMessagesToStream`
and `jsonmessage.DisplayJSONMessagesStream` functions completed
successfully or if the function returned early due the response object
being closed early.

For example, when using `http.NewRequestWithContext`, the request will
be cancelled when the context is cancelled. This usually closes the
`response` object, causing an `io.EOF`.

Signed-off-by: Alano Terblanche <18033717+Benehiko@users.noreply.github.com>